do while instead of while

+++ src/run.c
@@ -13,17 +13,15 @@ int run(char *cmd, char **env)
 {
 	int pid;
 
-	cmd = get_next_line(0);
-	while (my_strcmp(cmd, "exit") != 0) {
-		if (cmd == NULL)
-			return (84);
+	do {
 		pid = fork();
-		wait(NULL);
 		if (pid == 0)
 			exec_cmd(cmd, env);
-		else if (pid > 0)
+		else if (pid > 0 && wait(NULL) == 0)
 			kill(0, SIGKILL);
 		cmd = get_next_line(0);
-	}
+		if (cmd == NULL)
+			return (-1);
+	} while (my_strcmp(cmd, "exit") != 0);
 	return (pid);
 }
